Rename Sms{Service,Receiver} to WebOTPService
The old name does not reflect the changes made to the feature name
and specification. No change in functionality is expected.

The following changes were made using grep & sed.

- rename files
- [^_]+sms_service->webotp_service
- [^_]+sms_receiver->webotp_service
- {SMS,Sms}Receiver->WebOTPService
- SmsService->WebOTPService
- SMS Service->WebOTP Service

The only manual change was to update some outdated documentation in
webotp_service.mojom.

## Aside
Getting a pre-submit warning which seems incorrect as we have not
changes the UMA histogram name and it exists :-?

```
** Presubmit Warnings **
Some UMA_HISTOGRAM lines have been modified and the associated histogram name has no match in either tools/metrics/histograms/histograms.xml or the modifications of it:
   [components/browser_ui/sms/android/java/src/org/chromium/components/browser_ui/sms/WebOTPServiceUma.java:35] Blink.Sms.Receive.TimeCancelOnKeyboardDismissal
```

Bug:1117238

Change-Id: I369f8631450b0a539b3c2afbf8d4ed87ce853a87