Rework "SelfWillPaint-handling" in SVGContainerPainter
Check if the container has a filter effect property (with a reference
filter) before ensuring a paint chunk.
This both robust against paint-from-prepaint (crbug.com/848056) as well
as achieving the goal of the SelfWillPaint() check.
This also eliminates the last user of SelfWillPaint() so it and
SVGLayoutSupport::HasFilterResource() which it uses can be removed.

Bug: 109224, 848056, 1131498
Change-Id: Ic3da600c011d0a48bf562d9fea013e258d436e56